refactor items controllers
